Adding Rendezvous to support blocking ps ops
The Rendezvous gem from geemus was created to read the output of heroku
api commands which return immediately. This commit passes the attach:
true option to post_ps which makes the process return a rendezvous URL
that can be provided to Rendezvous to read from a socket to the URL and
print the output that it returns.  This has the effect of making the
rake db:migrations command block until the migrations have finished
running.